remove the obsolete DataVolume garbage collection check in validateUploadDataVolume, as it’s deprecated
